Add files via upload
Sales Data Analysis

Porject is to analyze the data and find the answer of question given below so that we can get best decision based on the this data analysis

What is the best month for sale?
Which city has max order?
What time should we display advertisements to maximise for product purchase?
What product sold the most? & Why?
What products are most often sold together?